pyproject.toml
[project]
name = "steam-review-tag-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.4.2",
    "ipykernel>=7.1.0",
    "ipython>=9.8.0",
    "mlflow>=3.8.1",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "files-to-prompt>=0.6",
]
